pyproject.toml
[project]
name = "cohere-explore"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.20.7",
    "python-dotenv>=1.2.1",
]
